glade clean build 완료

+++ payment-service/src/main/java/com/rushcrew/payment_service/application/command/PaymentCommand.java
@@ -5,6 +5,6 @@ import java.util.UUID;
 
 public record PaymentCommand(
         UUID orderId,
-        Long totalAmount
+        BigDecimal totalAmount
 ) {
 }
